Bugfix + Python venv
---

Bugfix: I need to check each element of expressions as a possible solution in each call to the recursive solver.

I found this because in the previous version of the code, this ...

target = 17
numbers = 3 10 7

... didn't result in 7 + 10 as a solution.

Clearly 7 + 10 = 17, but the old code would get to this point ...

expressions = [3, (17, +, 7, 10)]

... but then not see that the second element of the expressions list was a solution.

Doh!  Fixed now.

---

I also added support for a Python venv.
To use this tool ...

* Clone the repo
* Run ./install_requirements.sh
* Run ./countdown_numbers_solver_cli